+++ src/FsReveal/FsReveal.fs
@@ -8,205 +8,64 @@ open FSharp.Literate
 open FSharp.Markdown
 open FSharp.Markdown.Html
 
-type SlideData =
-  | Simple of MarkdownParagraph list
-  | Nested of MarkdownParagraph list list
+module FsRevealHelper = 
+    let mutable Folder = __SOURCE_DIRECTORY__
 
-type Slide = 
-  {
-    Properties: (string * string) list
-    SlideData: SlideData
-  }
-
-type Presentation = 
-  {
-    Properties: (string * string) list
-    Slides: Slide list
-    Document: LiterateDocument
-  }
-
-[<AutoOpen>]
-module internal Misc =
-
-  /// Correctly combine two paths
-  let (@@) a b = Path.Combine(a, b)
-
-  /// Ensure that directory exists
-  let ensureDirectory path =
-    let dir = DirectoryInfo(path)
-    if not dir.Exists then dir.Create()
-
-  /// Copy all files from source to target
-  let rec copyFiles source target =
-    ensureDirectory target
-    for f in Directory.GetDirectories(source) do
-      copyFiles f (target @@ Path.GetFileName(f))
-    for f in Directory.GetFiles(source) do
-      File.Copy(f, (target @@ Path.GetFileName(f)), true)
-
-  /// Split a list into chunks using the specified separator
-  /// This takes a list and returns a list of lists (chunks)
-  /// that represent individual groups, separated by the given
-  /// separator 'v'
-  let splitBy v list =
-    let yieldRevNonEmpty list = 
-      if list = [] then []
-      else [List.rev list]
-
-    let rec loop groupSoFar list = seq { 
-      match list with
-      | [] -> yield! yieldRevNonEmpty groupSoFar
-      | head::tail when head = v ->
-          yield! yieldRevNonEmpty groupSoFar
-          yield! loop [] tail
-      | head::tail ->
-          yield! loop (head::groupSoFar) tail }
-    loop [] list |> List.ofSeq
-
-
-  let getPresentation (doc:LiterateDocument) =
-    /// get properties, a list of (key,value) from
-    /// [[Span[Literal "key : value"]]]
-    let getProperties (spans: list<list<_>>) =
-      spans
-      |> List.map (fun c ->       
-        match c with
-        | [Span(l)] -> 
-          match l with
-          | [Literal(v)] when v.Contains(":") -> 
-              let colonPos = v.IndexOf(':')
-              let key = v.Substring(0, colonPos).Trim()
-              let value = v.Substring(colonPos + 1).Trim()
-              (key, value)
-          | _ -> failwithf "Invalid Presentation property: %A" l
-        | _ -> failwithf "Invalid Presentation property: %A" c
-      )
-  
-    // main section is separated by ***
-    let sections = splitBy (HorizontalRule('*')) doc.Paragraphs
-    let properties = 
-      match sections.Head with
-      | [ListBlock(_, spans)] -> getProperties spans
-      | x -> failwithf "Invalid Presentation properties: %A" x
-
-    let wrappedInSection properties paragraphs = 
-        let attributes = 
-            properties 
-            |> Seq.map (fun (k,v) -> sprintf "%s=\"%s\"" k v)
-            
-        InlineBlock(sprintf "<section %s>" (String.Join(" ",attributes))) :: paragraphs @ [InlineBlock("</section>")]
-
-    let getParagraphsFromSlide slide =
-      match slide.SlideData with
-      | Simple(paragraphs) ->
-          wrappedInSection slide.Properties paragraphs        
-      | Nested(listOfParagraphs) -> 
-          listOfParagraphs         
-          |> List.collect (wrappedInSection slide.Properties)
-          |> wrappedInSection slide.Properties
-
-    let slides = 
-      sections.Tail
-      |> List.map (fun s -> 
-          // sub-section is separated by ---
-          let result = splitBy (HorizontalRule('-')) s
-          let properties,data =
-              match s with
-              | ListBlock(_, spans) :: data -> 
-                  try
-                    getProperties spans,[data]
-                  with _ -> [],[s]
-              | _ -> [],[s]
-
-          {Properties = properties 
-           SlideData =
-              match data with
-              | [[slide]] -> Simple([slide])
-              | _ -> Nested(result) }
-        )
+type FsReveal = 
     
-    let paragraphs = List.collect (getParagraphsFromSlide) slides
+    static member GetPresentationFromScriptString fsx = 
+        let fsi = FsiEvaluator()
+        Literate.ParseScriptString(fsx, fsiEvaluator = fsi) |> getPresentation
+    
+    static member GetPresentationFromMarkdown md = 
+        md
+        |> Literate.ParseMarkdownString
+        |> getPresentation
+    
+    static member GenerateOutput outDir outFile presentation = 
+        if Directory.Exists outDir |> not then 
+            Directory.CreateDirectory outDir |> ignore
+            printfn "Creating %s.." outDir
+        let revealJsDir = FsRevealHelper.Folder @@ "../reveal.js"
+        printfn "Copy reveal.js files from %s to %s" revealJsDir outDir
+        copyFiles (fun f -> f.ToLower().Contains("index.html")) revealJsDir outDir
+        // delete overhead
+        File.Delete(outDir @@ "README.md")
+        let di = DirectoryInfo(outDir @@ "test")
+        if di.Exists then di.Delete(true)
+        let doc = Literate.FormatLiterateNodes presentation.Document
+        let htmlSlides = Literate.WriteHtml doc
+        let toolTips = doc.FormattedTips
+        let relative subdir = FsRevealHelper.Folder @@ subdir
+        printfn "Apply template : %s" (relative "template.html")
+        let output = StringBuilder(File.ReadAllText(relative "template.html"))
+        // replace properties
+        presentation.Properties |> List.iter (fun (k, v) -> output.Replace(sprintf "{%s}" k, v) |> ignore)
+        output.Replace("{slides}", htmlSlides).Replace("{tooltips}", toolTips) |> ignore
+        File.WriteAllText(outDir @@ outFile, output.ToString())
+    
+    static member private checkIfFileExistsAndRun file f = 
+        if File.Exists file then f()
+        else printfn "%s does not exist. Abort!" file
+    
+    static member GenerateOutputFromScriptFile outDir outFile fsxFile = 
+        FsReveal.checkIfFileExistsAndRun fsxFile (fun () -> 
+            fsxFile
+            |> File.ReadAllText
+            |> FsReveal.GetPresentationFromScriptString
+            |> FsReveal.GenerateOutput outDir outFile)
+    
+    static member GenerateOutputFromMarkdownFile outDir outFile mdFile = 
+        FsReveal.checkIfFileExistsAndRun mdFile (fun () -> 
+            mdFile
+            |> File.ReadAllText
+            |> FsReveal.GetPresentationFromMarkdown
+            |> FsReveal.GenerateOutput outDir outFile)
 
-    {
-      Properties = properties
-      Slides = slides
-      Document = doc.With(paragraphs = paragraphs)
-    }
-
-module FsRevealHelper =
-  let mutable Folder = __SOURCE_DIRECTORY__
-
-type FsReveal =
-  static member GetPresentationFromScriptString fsx =
-    let fsi = FsiEvaluator() 
-    Literate.ParseScriptString (fsx, fsiEvaluator = fsi)
-    |> getPresentation 
-
-  static member GetPresentationFromMarkdown md =
-    md 
-    |> Literate.ParseMarkdownString
-    |> getPresentation
-
-  static member GenerateOutput outDir outFile presentation =
-    if Directory.Exists outDir then
-      printfn "%s exists.." outDir
-    else
-      Directory.CreateDirectory outDir |> ignore
-      printfn "Create %s.." outDir
-
-    let revealJsDir = FsRevealHelper.Folder @@ "../reveal.js"
-    printfn "Copy reveal.js files from %s to %s" revealJsDir outDir 
-    copyFiles revealJsDir outDir
-
-    // delete overhead
-    File.Delete(outDir @@ "index.html")
-    File.Delete(outDir @@ "README.md")
-
-    let di = DirectoryInfo(outDir @@ "test")
-    di.Delete(true)
-
-    let doc = Literate.FormatLiterateNodes presentation.Document 
-  
-    let htmlSlides = Literate.WriteHtml doc
-    let toolTips = doc.FormattedTips
-
-    let relative subdir = FsRevealHelper.Folder @@ subdir    
-    printfn "Apply template : %s" (relative "template.html")
-    let output = StringBuilder(File.ReadAllText (relative "template.html"))    
-
-    // replace properties
-    presentation.Properties
-    |> List.iter (fun (k,v) -> 
-      let tag = sprintf "{%s}" k    
-      output.Replace(tag, v) |> ignore)
-
-    output
-      .Replace("{slides}", htmlSlides)
-      .Replace("{tooltips}", toolTips) |> ignore
-
-    File.WriteAllText (outDir @@ outFile, output.ToString())
-
-  
-  static member private checkIfFileExistsAndRun file f =
-    if File.Exists file then
-      f()
-    else
-      printfn "%s does not exist. Abort!" file
-
-  static member GenerateOutputFromScriptFile outDir outFile fsxFile =
-    FsReveal.checkIfFileExistsAndRun fsxFile (fun () -> 
-          let fsx = File.ReadAllText (fsxFile)
-
-          fsx 
-          |> FsReveal.GetPresentationFromScriptString
-          |> FsReveal.GenerateOutput outDir outFile
-      )
-
-  static member GenerateOutputFromMarkdownFile outDir outFile mdFile =
-    FsReveal.checkIfFileExistsAndRun mdFile (fun () -> 
-          let md = File.ReadAllText (mdFile)
-
-          md 
-          |> FsReveal.GetPresentationFromMarkdown
-          |> FsReveal.GenerateOutput outDir outFile
-      )       
+    static member GenerateFromFile outDir fileName =
+        let file = FileInfo fileName
+        let outputFileName = file.Name.Replace(file.Extension,".html")
+        match file.Extension with   
+        | ".md" -> FsReveal.GenerateOutputFromMarkdownFile outDir outputFileName file.FullName
+        | ".fsx" -> FsReveal.GenerateOutputFromScriptFile outDir outputFileName file.FullName
+        | _ -> ()