+++ tests/FsReveal.Tests/SectionsSpecs.fs
@@ -32,12 +32,13 @@ let md = """
 ### Slide 3"""
 
 [<Test>]
-let ``can generate sections from markdown`` () =
-  let slides = (md |> FsReveal.GetPresentationFromMarkdown).Slides
-
-  slides.Length |> shouldEqual 3
-  let slide = slides |> Seq.skip 1 |> Seq.head 
-  match slide.SlideData with
-  | SlideData.Nested x -> ()
-  | _ -> failwith "subslides not parsed"
- 
+let ``can generate sections from markdown``() = 
+    let slides = (md |> FsReveal.GetPresentationFromMarkdown).Slides
+    slides.Length |> shouldEqual 3
+    let slide = 
+        slides
+        |> Seq.skip 1
+        |> Seq.head
+    match slide.SlideData with
+    | SlideData.Nested x -> ()
+    | _ -> failwith "subslides not parsed"